LPS-89658 Sort

+++ modules/apps/headless/headless-web-experience/headless-web-experience-impl/src/main/java/com/liferay/headless/web/experience/internal/odata/entity/v1_0/EntityFieldsProvider.java
@@ -126,9 +126,9 @@ public class EntityFieldsProvider {
 					locale, "Number"));
 		}
 		else if (Objects.equals(
-					ddmFormField.getDataType(), FieldConstants.LONG) ||
+					ddmFormField.getDataType(), FieldConstants.INTEGER) ||
 				 Objects.equals(
-					 ddmFormField.getDataType(), FieldConstants.INTEGER)) {
+					 ddmFormField.getDataType(), FieldConstants.LONG)) {
 
 			return new IntegerEntityField(
 				_toFieldName(
